Added Classes & Object code written in Java
In Java, classes and objects are the fundamental building blocks of Object-Oriented Programming (OOP). They help in modeling real-world entities and their interactions in a program.

A class is a blueprint or template for creating objects. It defines the properties (fields/variables) and behaviors (methods/functions) that the objects created from the class will have.

An object is an instance of a class. It contains real values for the attributes defined in the class, and it can invoke methods.